move PlottingWorkspaceTreeItemState into PlottingViewState
