+++ qt/scientific_interfaces/ISISReflectometry/GUI/Plotting/presenter/PlottingViewStateBuilder.h
@@ -10,7 +10,6 @@
 #include "GUI/Common/PlotOptions.h"
 #include "GUI/Plotting/model/PlottingWorkspace.h"
 #include "GUI/Plotting/view/PlottingViewState.h"
-#include "GUI/Plotting/view/PlottingWorkspaceTreeItemState.h"
 
 #include <vector>
