+++ qt/scientific_interfaces/ISISReflectometry/GUI/Plotting/view/QtPlottingWorkspaceTreeViewAdapter.h
@@ -8,7 +8,7 @@
 
 #include "Common/DllConfig.h"
 #include "GUI/Plotting/model/PlottingWorkspace.h"
-#include "GUI/Plotting/view/PlottingWorkspaceTreeItemState.h"
+#include "GUI/Plotting/view/PlottingViewState.h"
 
 #include <QItemSelection>
 #include <QItemSelectionModel>
